tools/cc-gmail/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cc-gmail"
version = "0.1.0"
description = "Gmail CLI: read, send, search, and manage emails from the command line"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "CC Director Contributors"}]
requires-python = ">=3.11"
keywords = ["gmail", "email", "cli", "google", "api"]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "keyring>=25.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pyinstaller>=6.0.0"]

[project.scripts]
cc-gmail = "src.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
